requirements.txt
alice-images==0.1.0
apischema==0.15.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Flask==2.1.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kinopoisk-api-unofficial-client==2.2.0
lxml==4.8.0
MarkupSafe==2.1.1
pycparser==2.21
python-decouple==3.6
requests==2.26.0
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
